REMAINING: 3) Results check so the runs adds up executed vs failed 5) Renew "/tests" content with new name format "IO-MEM-Trapetzoidal_increase" 6) Prettify emulator help + \n in text 7) Distribution display in API is broken 8) Emulation logs are not there 9) Get the % value from total memory rather than from free memory 10) % in memory does not work 11) Rewrite CLI Help 15) Catch XML error handling in API 16) Add "dd" as emulator [way later] 17) Overlap check needs to be redone for multiple runs summary in carmelo distributions 18) Check why Trapetzoidal emulation NET -> log run wont start 19) Scheduler restart runs recovery
DONE:
#1) Remove distributions liear_decr, rev_trap, parabola, linear
#2) In restfully "/logs" have two "system" options
#4) "Granularity" field of the distribution is missing
#12) Change job key to "-j"
#13) Delete emulation by Name or ID
#14) Make stressaptes emulator IO load increase by file number - TEST NEEDED!!!!

+++ bin/EmulationManager.py
@@ -273,10 +273,7 @@ def deleteEmulation(emulationID):
     distributionName=[]
     
     try:
-        if HOMEPATH:
-            conn = sqlite.connect(HOMEPATH+'/data/cocoma.sqlite')
-        else:
-            conn = sqlite.connect('./data/cocoma.sqlite')
+        conn = sqlite.connect(HOMEPATH+'/data/cocoma.sqlite')
         c = conn.cursor()
         c.execute('SELECT distributionID,distributionName FROM distribution WHERE emulationID=?',[str(emulationID)])
                 
@@ -293,28 +290,60 @@ def deleteEmulation(emulationID):
                 c.execute('DELETE FROM DistributionParameters WHERE distributionID=?',[str(distributionID)])
                 c.execute('DELETE FROM EmulatorParameters WHERE distributionID=?',[str(distributionID)])
                 c.execute('DELETE FROM runLog WHERE distributionID=?',[str(distributionID)])
+            
+            c.execute('DELETE FROM distribution WHERE emulationID=?',[str(emulationID)])
+            c.execute('DELETE FROM emulationLifetime WHERE emulationID=?',[str(emulationID)])
+            c.execute('DELETE FROM emulation WHERE emulationID=?',[str(emulationID)])
+            
+            conn.commit()
+            c.close()
+            print "Emulation ID: ", emulationID," was deleted from DB"
                 
         else:
-            print "Emulation ID: "+str(emulationID)+" does not exists" 
-            return "Emulation ID: "+str(emulationID)+" does not exists"
-            sys.exit(1)
- 
-        c.execute('DELETE FROM distribution WHERE emulationID=?',[str(emulationID)])
-        c.execute('DELETE FROM emulationLifetime WHERE emulationID=?',[str(emulationID)])
-        c.execute('DELETE FROM emulation WHERE emulationID=?',[str(emulationID)])
-        #c.execute('DELETE FROM runLog WHERE distributionID=?',[str(distributionID)])
+            print "Emulation ID: "+str(emulationID)+" does not exists looking for name" 
+            c.execute('SELECT emulationID FROM emulation WHERE emulationName=?',[str(emulationID)])
+            emulationIDfetch = c.fetchall()
+            print emulationIDfetch
+            for row in emulationIDfetch:
+                emulationID = row[0]
+            c.execute('SELECT distributionID,distributionName FROM distribution WHERE emulationID=?',[str(emulationID)])
+                    
+            distributionIDfetch = c.fetchall()
+            
+            #getting list of distributions for emulation
+            if distributionIDfetch:
+                for row in distributionIDfetch:
+                    distributionID= row[0]
+                    distributionName.append(row[1])
+                    
+                    #deleting distribution related data
+                    c.execute('DELETE FROM DistributionParameters WHERE distributionID=?',[str(distributionID)])
+                    c.execute('DELETE FROM EmulatorParameters WHERE distributionID=?',[str(distributionID)])
+                    c.execute('DELETE FROM runLog WHERE distributionID=?',[str(distributionID)])
+                
+                c.execute('DELETE FROM distribution WHERE emulationID=?',[str(emulationID)])
+                c.execute('DELETE FROM emulationLifetime WHERE emulationID=?',[str(emulationID)])
+                c.execute('DELETE FROM emulation WHERE emulationID=?',[str(emulationID)])
+                    
+            else:
+                print "Emulation Name: "+str(emulationID)+" does not exists too" 
+                return "Emulation Name: "+str(emulationID)+" does not exists"
+                sys.exit(1)
+            
+            conn.commit()
+            c.close()
+            print "Emulation Name: ", emulationID," was deleted from DB"
         
-        
-        conn.commit()
+
     except sqlite.Error, e:
-        print "Could not delete emulationID: ",emulationID
+        print "Could not delete emulation: ",emulationID
         print "Error %s:" % e.args[0]
         print e
         return "Database error: "+str(e)
         sys.exit(1)
         
-    c.close()
-    print "Emulation ID: ", emulationID," was deleted from DB"
+    
+    
     
     #Now here we need to remove the emulation from the scheduler if exist
     uri ="PYRO:scheduler.daemon@"+str(readIfaceIP("schedinterface"))+":51889"
